fix(backend/[models,services]): change score field data types of AwardLeaderboard and CriteriaBreakdown to BigDecimal

+++ backend/backend/src/main/java/com/michaelcanonizado/backend/dtos/awardLeaderboard/AwardLeaderboardSummaryDTO.java
@@ -3,13 +3,14 @@ package com.michaelcanonizado.backend.dtos.awardLeaderboard;
 import com.michaelcanonizado.backend.dtos.candidate.CandidateSummaryDTO;
 import com.michaelcanonizado.backend.models.CriteriaBreakdown;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record AwardLeaderboardSummaryDTO(
         UUID id,
         CandidateSummaryDTO candidate,
-        Double score,
+        BigDecimal score,
         List<CriteriaBreakdown> criteriaBreakdown
 ) {
 }
